scraping service: added asynchronous calls to the scraping pages
