+++ infra/scraping-service/requirements.txt
@@ -7,3 +7,4 @@ pyppeteer
 selenium
 cloudscraper
 playwright
+aiohttp
