+++ infra/scraping-service/src/scraper/exceptions.py
@@ -1,0 +1,5 @@
+class HttpResponseException(Exception):
+    def __init__(self, message):
+        super().__init__(message)
+
+        self.message = message
