+++ infra/scraping-service/src/scraper/router.py
@@ -1,8 +1,9 @@
 from fastapi import APIRouter, Query
+from typing import Annotated
 
 from scraper.scraper import TrackerGGScraper
-from models.scraping_models import TrackerGGModel
-from typing import Annotated
+from models.models import TrackerGGModel
+
 
 router = APIRouter(
     prefix='/scrape'
@@ -16,8 +17,7 @@ async def scrape_data(model: Annotated[TrackerGGModel, Query()]):
         start_page=model.start_page,
         end_page=model.end_page, 
         platform=model.platform.value, 
-        game_mode=model.game_mode.value,
-        write_to_file=False
+        game_mode=model.game_mode.value
     )
 
 @router.post('')
@@ -28,6 +28,5 @@ async def scrape_data_with_write(model: Annotated[TrackerGGModel, Query()]):
         start_page=model.start_page,
         end_page=model.end_page, 
         platform=model.platform.value, 
-        game_mode=model.game_mode.value,
-        write_to_file=True
+        game_mode=model.game_mode.value
     )
